Gitignore server build files
Removing code that is compiled and produced through the webpack build process from version control. No need to store the content in the node_modules directory as it can be reproduced with npm. The same goes for the server build files as it will be reproduced with webpack. This will be done locally with only the source code and compilation process being tracked under version control